Remove default hostname when regenerating pki and changed to a prompt
The `zarf pki regenerate` listed 'zarf-server' as the default host but would also
require the user to pass the --host flag anyways. This was counterintuitive. Updated
so that if host is not passed as a flag the user would be prompted, similar to how
it works in the init command.

Signed-off-by: Jeff McCoy <code@jeffm.us>