HISTORY REWRITE.  Fixes #53, sign-off all existing commits
This commit is the completion of a history rewrite to sign-off all unsigned commits in the repo from its inception.  Please see the comments on https://github.com/defenseunicorns/zarf/issues/53 for more details on how to move your WIP if you have a branch you are working with commits ahead of master.

Note that because this is a rewrite pulling master will return an error the first time unless you run `git pull --force`.  Make sure you stash or commit your changes if you have anything on master locally that isn't merged already.