prevent breaking config change, rename componentName->name
To prevent the json schema from breaking with prior versions, making the name key optional still allows custom name mappings, but defaults to just use the same name as the main component.

Also updated the json schema and shortened the name key since we are already pointing to another component.