Switch to distribution/distribution vs regex hacking for image url pa… (#876)
This PR removes the Frankenstein regex we were using for parsing docker
image references and instead uses the same code that the docker registry
itself uses for parsing references to handle various edges cases. SHA1
hashes are also replaced with CRC32 for images and git URLs.

The image replacement also should support image digests, but more work
will need to be done on the usage of crane for image push/pull as
digests are not yet preserved.