Fix the double ui build issues (#894)
## Description

A bug was introduces to `make build-cli` that builds the UI twice. This
PR fixes that.

## Related Issue

Fixes #N/A

## Type of change

- [X] Bug fix (non-breaking change which fixes an issue)

## Checklist before merging


- [X] (Optional) Changes have been linted locally with
[golangci-lint](https://github.com/golangci/golangci-lint). (NOTE: We
haven't turned on lint checks in the pipeline yet so linting may be hard
if it shows a lot of lint errors in places that weren't touched by
changes. Thus, linting is optional right now.)

Co-authored-by: Megamind <882485+jeff-mccoy@users.noreply.github.com>