change enable --now to enable and start (#953)
## Description

Systemd versions before 2015 (which are still supported under 10 yr
extended maintenance) do not respect the --now flag. See
https://github.com/systemd/systemd/blob/dd95b381b26369c02a3ca03585184cb2ea17295b/NEWS#L3043-L3046

## Related Issue

Fixes #N/A

## Type of change

- [X] Bug fix (non-breaking change which fixes an issue)

## Checklist before merging

- [X] (Optional) Changes have been linted locally with
[golangci-lint](https://github.com/golangci/golangci-lint). (NOTE: We
haven't turned on lint checks in the pipeline yet so linting may be hard
if it shows a lot of lint errors in places that weren't touched by
changes. Thus, linting is optional right now.)