Use a regex manager to look at zarf.yamls and PR for their helm charts (#963)
## Description

This adds a regex manager for looking inside of zarf.yaml files and
pulling out helm charts. This will be useful to know when gitea (or any
other charts we add) are out of date and in need of updating.

## Related Issue

Fixes # N/A

## Type of change

- [X] New feature (non-breaking change which adds functionality)

## Checklist before merging

- [X] (Optional) Changes have been linted locally with
[golangci-lint](https://github.com/golangci/golangci-lint). (NOTE: We
haven't turned on lint checks in the pipeline yet so linting may be hard
if it shows a lot of lint errors in places that weren't touched by
changes. Thus, linting is optional right now.)

Co-authored-by: Megamind <882485+jeff-mccoy@users.noreply.github.com>