Support automatic multi-part packages based on a MaxPackageSize (#848)
This PR adds the `zarf package create --max-package-size=N`, where N is
the maximum number of megabytes a part can be, with the default being 0.
Above the size limit, a Zarf package is automatically broken into
smaller chunks up to the limit. These chunks, starting with index 000,
are then reassembled into a single package on `zarf package deploy`
before running the deployment. This means that if a user attempts to
deploy a multi-part Zarf package but does not confirm the deployment,
the multipart archive will now be just a single archive still. This was
needed to avoid putting 3 copies of the data on the disk (the original
multi-part, the assembled archive, and the temp folder where the archive
is extracted).

Fixes #843 

## Type of change
- [x] New feature (non-breaking change which adds functionality)
- [x] Breaking change (fix or feature that would cause existing
functionality to not work as expected)