Expose PVC accessMode as variable (#1664)
## Description

When attempting to upgrade our EKS Node AMIs in AWS, we noticed that the
Zarf Registry deployment was unable to horizontally scale across nodes
which needed to restart. We believe the culprit is the `accessMode`
specification for the PersistentVolumeController. In order for multiple
pods to have access to the same PersistentVolume, the `accessMode` must
be set to "ReadWriteMany".

~~This PR proposes that when autoscaling is enabled for the Zarf
Registry, the `accessMode` is set to "ReadWriteMany" by default; when
autoscaling is disabled, it is set to "ReadWriteOnce". Due to the
additional work required (i.e. using an existing
PersistencVolumeController with a storage class compatible with RWX), we
also propose that `autoscaling` be disabled by default.~~

This PR exposes the `REGISTRY_PVC_ACCESS_MODE` variable for the
`zarf-registry` portion of the init package.

## Related Issue

- Relates to #375 

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Other (security config, docs update, etc)

## Checklist before merging

- [x] Test, docs, adr added or updated as needed
- [x] [Contributor Guide
Steps](https://github.com/defenseunicorns/zarf/blob/main/CONTRIBUTING.md#developer-workflow)
followed