Adding crane ls command (#1835)
## Description

This adds support for `crane ls`. This feature is very useful when
attempting to do some automation of updating images. The only other way
to do this is to run `zarf connect registry` and then request the tags
endpoint of the registry.

## Related Issue

N/A

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Other (security config, docs update, etc)

## Checklist before merging

- [ ] Test, docs, adr added or updated as needed
- [x] [Contributor Guide
Steps](https://github.com/defenseunicorns/zarf/blob/main/CONTRIBUTING.md#developer-workflow)
followed