Refactors and library additions to support `bundle`s (#1770)
## Description

This PR performs some refactoring in `packager`, `oci` and a few other
places to allow for orchestration of Zarf packages in `bundles`. A POC
of this bundle effort is/was contained in this PR, but will/has been
migrated to a new CLI, and as such been removed from this PR. The final
state of this PR will allow for easier consumption of certain
functionality in `packager`, more powerful `oci` library features, and
is one of the first true tests of Zarf as a library. 2 months well
spent.

Closes https://github.com/defenseunicorns/zarf/issues/1908

## Type of change

- [x] New feature (non-breaking change which adds functionality)

## Checklist before merging

- [x] Test, docs, adr added or updated as needed
- [x] [Contributor Guide
Steps](https://github.com/defenseunicorns/zarf/blob/main/CONTRIBUTING.md#developer-workflow)
followed

---------

Signed-off-by: razzle <harry@razzle.cloud>
Co-authored-by: Wayne Starr <Racer159@users.noreply.github.com>
Co-authored-by: UncleGedd <42304551+UncleGedd@users.noreply.github.com>