adds a hint when the release might be empty (#2003)
## Description

This assists when you may include a `bigbang` package but for some
reason a dependency is not included. A good example of this is when you
try to include `mattermost` and `mattermostOperator` but you forget, or
don't know to include `minioOperator`. The extension attempts to install
it, but doesn't include the name of what it's trying to install. This
gives a hint by including the namespace in the output. There might be a
better way to solve this, but this seems to help me figure out what to
do next.

Without this the output is:

```
Big Bang Helm Release `` to be ready
```

With this change the output is:

```
Big Bang Helm Release `` to be ready (Namespace: bigbang.minio-operator)
```

## Related Issue

N/A

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Other (security config, docs update, etc)

## Checklist before merging

- [x] Test, docs, adr added or updated as needed
- [x] [Contributor Guide
Steps](https://github.com/defenseunicorns/zarf/blob/main/CONTRIBUTING.md#developer-workflow)
followed

---------

Co-authored-by: Wayne Starr <Racer159@users.noreply.github.com>