Add a Loader interface for loading package definitions on create
"Normal" packages and skeleton packages require different
loading processes, so using an interface allows us to
have two separate implementations for them.

This makes the load logic easier to reason about and easier to maintain.

Loader is defined in package packager because most of the operations
require direct access to the top-level Packager struct fields.