fix: allow usage of local image archives again
a check in create_stages during package creation tries to
parses references for the images provided for components in
the zarf.yaml config. By default the reference parsing assumes
that the source of the image is that of a container registry.
If the image is a tarball instead, the reference parser errors out.

Check if the image reference is that of a tarball and skips the
reference check if it is so allowing the image to be loaded by the cli.

Signed-off-by: Vibhav Bobade <vibhav.bobde@gmail.com>