fix: allow host+subpath as the source registry for registry-override (#2135)
Instead of looking for refInfo.Host in the override map loop through the keys and values in i.RegistryOverrides, check if the refInfo.Reference begins with an override key and, if it does, replace that override text with the override value and set it back to actualSrc.

Do not use ImageTransformHostWithoutChecksum since we already have the parsed ref and all the info we need to do the replacement.

default actualSrc to refInfo.Reference

Signed-off-by: Vibhav Bobade <vibhav.bobde@gmail.com>