fix: hotfix skeleton publish (#2398)
## Description

Hotfixes skeleton publish. However, this breakage has revealed that the
entire `skeleton` creation + publishing flow is flawed and requires an
overhaul. Currently, while `skeleton` says it has an `arch` of
`skeleton`, it is actually honoring the `--architecture` flag and
resulting in a `skeleton` package that is single architecture. It is
also not honoring `flavor`.

## Type of change

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Other (security config, docs update, etc)

---------

Signed-off-by: razzle <harry@razzle.cloud>