ci: run revive using golang-lint-ci (#2499)
## Description

This replaces revive with golang-ci-lint, and instead runs revive
through the aforementioned tool. Additionally it replaces the header
check script with the goheader linter. The benefit of this change is
that we can enable more linters along with revive in the future.

## Related Issue

Part of #2503 

## Checklist before merging

- [x] Test, docs, adr added or updated as needed
- [x] [Contributor Guide
Steps](https://github.com/defenseunicorns/zarf/blob/main/.github/CONTRIBUTING.md#developer-workflow)
followed

---------

Co-authored-by: Austin Abro <37223396+AustinAbro321@users.noreply.github.com>
Co-authored-by: razzle <harry@razzle.cloud>