test: pod agent unit tests (#2526)
## Description

This also changes how we the agent gets secrets from the cluster.
Instead of mounting secrets in the deployment and reading from a file
the agent will instead read the secret directly from the cluster. This
is done for simplicity, and easier testing. Additionally, the agent will
need to read from the cluster in #1974 as well so this helps prepare for
that change.

This only changes and tests pods to keep the scope of the PR slim. This
means temporarily the agent will both read from mounted file secrets and
regular kubernetes secrets depending on the hook, but this will be
changed shortly in future PRs.

## Related Issue

Relates to #2512 

## Checklist before merging

- [ ] Test, docs, adr added or updated as needed
- [ ] [Contributor Guide
Steps](https://github.com/defenseunicorns/zarf/blob/main/.github/CONTRIBUTING.md#developer-workflow)
followed

---------

Co-authored-by: Lucas Rodriguez <lucas.rodriguez@defenseunicorns.com>