fix: crane option argument parameters (#2609)
## Description

During refactoring in #2460 we broke the option parameters for Crane.
Honestly I am not surprised that this has happened. The way that these
options are parsed and passed around are a threading mess. The issue is
that this seems to also be the method in which Crane does it. This
change reverts the specific pointer changes which caused the issue in
the first place.

This is one the reasons why I do not like Cobra. The use of init
functions and global variables are doomed to cause threading issues.
Reminds me of writing Java programs with Spring Boot.

## Related Issue

Fixes #2571 

## Checklist before merging

- [x] Test, docs, adr added or updated as needed
- [x] [Contributor Guide
Steps](https://github.com/defenseunicorns/zarf/blob/main/.github/CONTRIBUTING.md#developer-workflow)
followed

Co-authored-by: Austin Abro <37223396+AustinAbro321@users.noreply.github.com>