feat: make the registry overrides deterministic and ordered
Provides an ordered, structured mechanism to specify registry overrides.
A helper function is provided.

For context, the current implementation does not allow overriding a
repository within a registry in a reliable manner. One example is
docker.io and docker.io/library. To map both of these to an internal
repository, (eg. docker.example.com/third-party), the distinction
between `/library` and `/unofficial-image` must be available.

By internally representing the overrides as an array, we maintain the
order the user originally specified. This allows a user to leverage a
longest-prefix-first ordering to override registry repositories in a
predictable manner.

Signed-off-by: Richard Luby <richluby@users.noreply.github.com>